# swagger.yaml
openapi: 3.0.0
info:
  title: Blog API
  version: 1.0.0
  description: A simple API for managing blog posts.

servers:
  - url: http://localhost:3000/api

paths:
  /posts:
    post:
      summary: Create a new blog post
      description: Creates a blog post with a title and content.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "My First Post"
                content:
                  type: string
                  example: "This is the content of my first post."
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  content:
                    type: string

    get:
      summary: Get all blog posts
      responses:
        '200':
          description: A list of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    content:
                      type: string

  /posts/{id}:
    get:
      summary: Get a blog post by ID
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the blog post
          schema:
            type: integer
      responses:
        '200':
          description: Blog post found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  content:
                    type: string
        '404':
          description: Post not found

    put:
      summary: Update a blog post
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the blog post
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated Title"
                content:
                  type: string
                  example: "Updated content."
      responses:
        '200':
          description: Blog post updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  content:
                    type: string
        '404':
          description: Post not found

    delete:
      summary: Delete a blog post
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the blog post
          schema:
            type: integer
      responses:
        '204':
          description: Post deleted successfully
        '404':
          description: Post not found